third_party/WebKit/LayoutTests/fast/dom/DeviceLight/add-listener-expected.txt

Tests that adding a new devicelight event listener invokes a callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deviceLightEvent.value is mockLight
PASS successfullyParsed is true

TEST COMPLETE
